salt states: split cordon from kubernetes monolithic file
